+++ cmplx/polar.go
@@ -7,6 +7,6 @@ package cmplx
 // Polar returns the absolute value r and phase θ of x,
 // such that x = r * e**θi.
 // The phase is in the range [-Pi, Pi].
-func Polar(x complex128) (r, θ float64) {
+func Polar(x complex128) (r, θ float32) {
 	return Abs(x), Phase(x)
 }
